+++ .gitignore
@@ -9,6 +9,9 @@
 /node_modules/
 /npm-debug.log
 
+## failed screenshots
+/screenshots/*/failed/
+
 ## testing
 /coverage/
