+++ demo/index.html
@@ -17,7 +17,7 @@
 
     <script type="module">
       import { html, render } from 'lit';
-      import '../dist/src/oscd-filtered-list.js';
+      import '../dist/OscdFilteredList.js';
 
       const list = [
         { prim: 'item1prim', sec: 'item1sec', disabled: false },
