+++ tsconfig.json
@@ -10,14 +10,13 @@
     "allowSyntheticDefaultImports": true,
     "experimentalDecorators": true,
     "importHelpers": true,
-    "allowJs": true,
     "skipLibCheck": true,
     "outDir": "dist",
     "sourceMap": true,
     "inlineSources": true,
-    "rootDir": "./",
+    "rootDir": "./src",
     "declaration": true,
     "incremental": true
   },
-  "include": ["**/*.ts"]
+  "include": ["src/*.ts"]
 }
